third_party/WebKit/LayoutTests/editing/inserting/insert_div_before_br_at_end.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    '<div contenteditable>one\u{00A0}|<br></div>',
    'insertHTML two<div>three</div>',
    '<div contenteditable>one two<div>three|</div></div>'),
    'insert DIV before BR at end');
</script>
